Refactored Employee auth to use link instead of OTP for email confirmation and reset password.--Made some changes to menu, dish, and employee models.-- Modified email template files.-- Modifed email messages.-- Implemented another method to handle email sending.

+++ controllers/menuController.js
@@ -60,6 +60,7 @@ const remove = async (req, res) => {
     try {
             const menu = await Menu.findById({ _id: id });
             if(!menu) return res.status(404).json({message: `No menu record found for Id: ${id}`});
+            if(menu.dishes) return res.status(400).json({message: 'You can not delete a menu that already has a list of dishes.'});
 
             await Menu.deleteOne({ _id: id });
             res.status(200).json({message: 'Menu successfully deleted'});
